backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pillow==11.0.0
pydantic==2.10.3
python-multipart